+++ src/comparator.js
@@ -1,11 +1,25 @@
-// utils
-import {areIterablesEqual, isPromiseLike, isReactElement, sameValueZeroEqual} from './utils';
+// constants
+import {HAS_MAP_SUPPORT, HAS_SET_SUPPORT} from './constants';
 
-const HAS_MAP_SUPPORT = typeof Map === 'function';
-const HAS_SET_SUPPORT = typeof Set === 'function';
+// utils
+import {
+  areArraysEqual,
+  areObjectsEqual,
+  createAreIterablesEqual,
+  isPlainObject,
+  isPromiseLike,
+  isReactElement,
+  sameValueZeroEqual
+} from './utils';
+
+const isArray = Array.isArray;
+
+const areMapsEqual = createAreIterablesEqual(true);
+const areSetsEqual = createAreIterablesEqual(false);
 
 const createComparator = (createIsEqual) => {
-  const isEqual = typeof createIsEqual === 'function' ? createIsEqual(comparator) : comparator; // eslint-disable-line
+  // eslint-disable-next-line no-use-before-define
+  const isEqual = typeof createIsEqual === 'function' ? createIsEqual(comparator) : comparator;
 
   /**
    * @function comparator
@@ -13,113 +27,77 @@ const createComparator = (createIsEqual) => {
    * @description
    * compare the value of the two objects and return true if they are equivalent in values
    *
-   * @param {*} objectA the object to test against
-   * @param {*} objectB the object to test
+   * @param {any} objectA the object to test against
+   * @param {any} objectB the object to test
+   * @param {any} [meta] an optional meta object that is passed through to all equality test calls
    * @returns {boolean} are objectA and objectB equivalent in value
    */
-  function comparator(objectA, objectB) {
+  function comparator(objectA, objectB, meta) {
     if (sameValueZeroEqual(objectA, objectB)) {
       return true;
     }
 
     const typeOfA = typeof objectA;
 
-    if (typeOfA !== typeof objectB) {
+    if (typeOfA !== typeof objectB || typeOfA !== 'object' || !objectA || !objectB) {
       return false;
     }
 
-    if (typeOfA === 'object' && objectA && objectB) {
-      const arrayA = Array.isArray(objectA);
-      const arrayB = Array.isArray(objectB);
-
-      let index;
-
-      if (arrayA || arrayB) {
-        if (arrayA !== arrayB || objectA.length !== objectB.length) {
-          return false;
-        }
-
-        for (index = 0; index < objectA.length; index++) {
-          if (!isEqual(objectA[index], objectB[index])) {
-            return false;
-          }
-        }
-
-        return true;
-      }
-
-      const dateA = objectA instanceof Date;
-      const dateB = objectB instanceof Date;
-
-      if (dateA || dateB) {
-        return dateA === dateB && sameValueZeroEqual(objectA.getTime(), objectB.getTime());
-      }
-
-      const regexpA = objectA instanceof RegExp;
-      const regexpB = objectB instanceof RegExp;
-
-      if (regexpA || regexpB) {
-        return (
-          regexpA === regexpB &&
-          objectA.source === objectB.source &&
-          objectA.global === objectB.global &&
-          objectA.ignoreCase === objectB.ignoreCase &&
-          objectA.multiline === objectB.multiline &&
-          objectA.lastIndex === objectB.lastIndex
-        );
-      }
-
-      if (isPromiseLike(objectA) || isPromiseLike(objectB)) {
-        return objectA === objectB;
-      }
-
-      if (HAS_MAP_SUPPORT) {
-        const mapA = objectA instanceof Map;
-        const mapB = objectB instanceof Map;
-
-        if (mapA || mapB) {
-          return mapA === mapB && areIterablesEqual(objectA, objectB, comparator, true);
-        }
-      }
-
-      if (HAS_SET_SUPPORT) {
-        const setA = objectA instanceof Set;
-        const setB = objectB instanceof Set;
-
-        if (setA || setB) {
-          return setA === setB && areIterablesEqual(objectA, objectB, comparator, false);
-        }
-      }
-
-      const keysA = Object.keys(objectA);
-
-      if (keysA.length !== Object.keys(objectB).length) {
-        return false;
-      }
-
-      let key;
-
-      for (index = 0; index < keysA.length; index++) {
-        key = keysA[index];
-
-        if (!Object.prototype.hasOwnProperty.call(objectB, key)) {
-          return false;
-        }
-
-        // if a react element, ignore the "_owner" key because its not necessary for equality comparisons
-        if (key === '_owner' && isReactElement(objectA) && isReactElement(objectB)) {
-          continue;
-        }
-
-        if (!isEqual(objectA[key], objectB[key])) {
-          return false;
-        }
-      }
-
-      return true;
+    if (isPlainObject(objectA) && isPlainObject(objectB)) {
+      return areObjectsEqual(objectA, objectB, isEqual, meta);
     }
 
-    return false;
+    const arrayA = isArray(objectA);
+    const arrayB = isArray(objectB);
+
+    if (arrayA || arrayB) {
+      return arrayA === arrayB && areArraysEqual(objectA, objectB, isEqual, meta);
+    }
+
+    const dateA = objectA instanceof Date;
+    const dateB = objectB instanceof Date;
+
+    if (dateA || dateB) {
+      return dateA === dateB && sameValueZeroEqual(objectA.getTime(), objectB.getTime());
+    }
+
+    const regexpA = objectA instanceof RegExp;
+    const regexpB = objectB instanceof RegExp;
+
+    if (regexpA || regexpB) {
+      return (
+        regexpA === regexpB &&
+        objectA.source === objectB.source &&
+        objectA.global === objectB.global &&
+        objectA.ignoreCase === objectB.ignoreCase &&
+        objectA.multiline === objectB.multiline &&
+        objectA.lastIndex === objectB.lastIndex
+      );
+    }
+
+    if (isPromiseLike(objectA) || isPromiseLike(objectB)) {
+      return objectA === objectB;
+    }
+
+    if (HAS_MAP_SUPPORT) {
+      const mapA = objectA instanceof Map;
+      const mapB = objectB instanceof Map;
+
+      if (mapA || mapB) {
+        return mapA === mapB && areMapsEqual(objectA, objectB, comparator, meta);
+      }
+    }
+
+    if (HAS_SET_SUPPORT) {
+      const setA = objectA instanceof Set;
+      const setB = objectB instanceof Set;
+
+      if (setA || setB) {
+        return setA === setB && areSetsEqual(objectA, objectB, comparator, meta);
+      }
+    }
+
+    return areObjectsEqual(objectA, objectB, isEqual, meta);
   }
 
   return comparator;