+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # fast-equals CHANGELOG
 
+## 1.5.0
+
+* Add [`circularDeepEqual`](README.md#circulardeepequal) and [`circularShallowEqual`](README.md#circularshallowequal) methods
+* Add `meta` third parameter to `comparator` calls, for use with `createCustomEqual` method
+
 ## 1.4.1
 
 * Fix issue where `lastIndex` was not being tested on `RegExp` objects